Move headers models into their own module (#7647)
