+++ apollo-federation/src/connectors/models/headers.rs
@@ -1,0 +1,201 @@
+#![deny(clippy::pedantic)]
+
+use std::error::Error;
+use std::fmt::Display;
+use std::fmt::Formatter;
+
+use apollo_compiler::Node;
+use apollo_compiler::ast;
+use apollo_compiler::parser::SourceSpan;
+use either::Either;
+use http::HeaderName;
+use http::header;
+
+use crate::connectors::JSONSelection;
+use crate::connectors::header::HeaderValue;
+use crate::connectors::spec::schema::HEADERS_ARGUMENT_NAME;
+use crate::connectors::spec::schema::HTTP_HEADER_MAPPING_FROM_ARGUMENT_NAME;
+use crate::connectors::spec::schema::HTTP_HEADER_MAPPING_NAME_ARGUMENT_NAME;
+use crate::connectors::spec::schema::HTTP_HEADER_MAPPING_VALUE_ARGUMENT_NAME;
+use crate::connectors::string_template;
+
+#[derive(Clone, Debug)]
+pub(crate) struct Header<'a> {
+    pub(crate) name: HeaderName,
+    pub(crate) name_node: &'a Node<ast::Value>,
+    pub(crate) source: HeaderSource,
+    pub(crate) source_node: &'a Node<ast::Value>,
+}
+
+impl<'a> Header<'a> {
+    /// Get a list of headers from the `headers` argument in a `@connect` or `@source` directive.
+    pub(crate) fn from_headers_arg(
+        node: &'a Node<ast::Value>,
+    ) -> Vec<Result<Self, HeaderParseError<'a>>> {
+        match (node.as_list(), node.as_object()) {
+            (Some(values), _) => values.iter().map(Self::from_single).collect(),
+            (None, Some(_)) => vec![Self::from_single(node)],
+            _ => vec![Err(HeaderParseError::Other {
+                message: format!("`{HEADERS_ARGUMENT_NAME}` must be an object or list of objects"),
+                node,
+            })],
+        }
+    }
+
+    /// Build a single [`Self`] from a single entry in the `headers` arg.
+    fn from_single(node: &'a Node<ast::Value>) -> Result<Self, HeaderParseError<'a>> {
+        let mappings = node.as_object().ok_or_else(|| HeaderParseError::Other {
+            message: "the HTTP header mapping is not an object".to_string(),
+            node,
+        })?;
+        let name_node = mappings
+            .iter()
+            .find_map(|(name, value)| {
+                (*name == HTTP_HEADER_MAPPING_NAME_ARGUMENT_NAME).then_some(value)
+            })
+            .ok_or_else(|| HeaderParseError::Other {
+                message: format!("missing `{HTTP_HEADER_MAPPING_NAME_ARGUMENT_NAME}` field"),
+                node,
+            })?;
+        let name = name_node
+            .as_str()
+            .ok_or_else(|| format!("`{HTTP_HEADER_MAPPING_NAME_ARGUMENT_NAME}` is not a string"))
+            .and_then(|name_str| {
+                HeaderName::try_from(name_str)
+                    .map_err(|_| format!("the value `{name_str}` is an invalid HTTP header name"))
+            })
+            .map_err(|message| HeaderParseError::Other {
+                message,
+                node: name_node,
+            })?;
+
+        if RESERVED_HEADERS.contains(&name) {
+            return Err(HeaderParseError::Other {
+                message: format!("header '{name}' is reserved and cannot be set by a connector"),
+                node: name_node,
+            });
+        }
+
+        let from = mappings
+            .iter()
+            .find(|(name, _value)| *name == HTTP_HEADER_MAPPING_FROM_ARGUMENT_NAME);
+        let value = mappings
+            .iter()
+            .find(|(name, _value)| *name == HTTP_HEADER_MAPPING_VALUE_ARGUMENT_NAME);
+
+        match (from, value) {
+            (Some(_), None) if STATIC_HEADERS.contains(&name) => {
+                Err(HeaderParseError::Other{ message: format!(
+                    "header '{name}' can't be set with `{HTTP_HEADER_MAPPING_FROM_ARGUMENT_NAME}`, only with `{HTTP_HEADER_MAPPING_VALUE_ARGUMENT_NAME}`"
+                ), node: name_node})
+            }
+            (Some((_, from_node)), None) => {
+                from_node.as_str()
+                    .ok_or_else(|| format!("`{HTTP_HEADER_MAPPING_FROM_ARGUMENT_NAME}` is not a string"))
+                    .and_then(|from_str| {
+                        HeaderName::try_from(from_str).map_err(|_| {
+                            format!("the value `{from_str}` is an invalid HTTP header name")
+                        })
+                    })
+                    .map(|from| Self {
+                        name,
+                        name_node,
+                        source: HeaderSource::From(from),
+                        source_node: from_node,
+                    })
+                    .map_err(|message| HeaderParseError::Other{ message, node: from_node})
+            }
+            (None, Some((_, value_node))) => {
+                value_node
+                    .as_str()
+                    .ok_or_else(|| HeaderParseError::Other{ message: format!("`{HTTP_HEADER_MAPPING_VALUE_ARGUMENT_NAME}` field in HTTP header mapping must be a string"), node: value_node})
+                    .and_then(|value_str| {
+                        value_str
+                            .parse::<HeaderValue>()
+                            .map_err(|err| HeaderParseError::ValueError {err, node: value_node})
+                    })
+                    .map(|value| Self {
+                        name,
+                        name_node,
+                        source: HeaderSource::Value(value),
+                        source_node: value_node,
+                    })
+            }
+            (None, None) => {
+                Err(HeaderParseError::Other {
+                    message: format!("either `{HTTP_HEADER_MAPPING_FROM_ARGUMENT_NAME}` or `{HTTP_HEADER_MAPPING_VALUE_ARGUMENT_NAME}` must be set"),
+                    node,
+                })
+            },
+            (Some((from_name, _)), Some((value_name, _))) => {
+                Err(HeaderParseError::ConflictingArguments {
+                    message: format!("`{HTTP_HEADER_MAPPING_FROM_ARGUMENT_NAME}` and `{HTTP_HEADER_MAPPING_VALUE_ARGUMENT_NAME}` can't be set at the same time"),
+                    from_location: from_name.location(),
+                    value_location: value_name.location(),
+                })
+            }
+        }
+    }
+}
+
+#[derive(Clone, Debug)]
+pub enum HeaderSource {
+    From(HeaderName),
+    Value(HeaderValue),
+}
+
+impl HeaderSource {
+    pub(crate) fn expressions(&self) -> impl Iterator<Item = &JSONSelection> {
+        match self {
+            HeaderSource::From(_) => Either::Left(std::iter::empty()),
+            HeaderSource::Value(value) => Either::Right(value.expressions().map(|e| &e.expression)),
+        }
+    }
+}
+
+#[derive(Debug)]
+pub(crate) enum HeaderParseError<'a> {
+    ValueError {
+        err: string_template::Error,
+        node: &'a Node<ast::Value>,
+    },
+    /// Both `value` and `from` are set
+    ConflictingArguments {
+        message: String,
+        from_location: Option<SourceSpan>,
+        value_location: Option<SourceSpan>,
+    },
+    Other {
+        message: String,
+        node: &'a Node<ast::Value>,
+    },
+}
+
+impl Display for HeaderParseError<'_> {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Self::ConflictingArguments { message, .. } | Self::Other { message, .. } => {
+                write!(f, "{message}")
+            }
+            Self::ValueError { err, .. } => write!(f, "{err}"),
+        }
+    }
+}
+
+impl Error for HeaderParseError<'_> {}
+
+const RESERVED_HEADERS: [HeaderName; 11] = [
+    header::CONNECTION,
+    header::PROXY_AUTHENTICATE,
+    header::PROXY_AUTHORIZATION,
+    header::TE,
+    header::TRAILER,
+    header::TRANSFER_ENCODING,
+    header::UPGRADE,
+    header::CONTENT_LENGTH,
+    header::CONTENT_ENCODING,
+    header::ACCEPT_ENCODING,
+    HeaderName::from_static("keep-alive"),
+];
+
+const STATIC_HEADERS: [HeaderName; 3] = [header::CONTENT_TYPE, header::ACCEPT, header::HOST];
